pyproject.toml
[tool.poetry]
name = "fastapi-authenticated"
version = "0.1.0"
description = ""
authors = ["harokki"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
taskipy = "^1.8.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
dependency-injector = "^4.36.0"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.8b0"
flake8 = "^3.9.2"
pytest = "^6.2.5"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = { cmd = "uvicorn app.main:app --reload", help = "launch test server" }
